chat/requirements.txt
streamlit
speechrecognition
gtts
sqlite3
google-generativeai